fix PR points and made UI dynamic of statitics screen
