+++ src/component/ProgressBar.js
@@ -1,32 +1,42 @@
 import React from 'react';
 import { View, Text, StyleSheet, Dimensions, ScrollView } from 'react-native';
 import * as Progress from 'react-native-progress';
+import styles from '../styles/ProgressStyle';
 
-const categories = [
-    { label: 'Shopping', value: 0.8, color: '#FCAC12', amount: '5120' },
-    { label: 'Subscription', value: 0.2, color: '#6A00FF', amount: '1280' },
-    { label: 'Food', value: 0.1, color: '#FD3C4A', amount: '532' },
-    { label: 'Grocery', value: 0.1, color: '#00A86B', amount: '412' },
-    { label: 'Food', value: 0.1, color: '#FD3C4A', amount: '532' },
-    { label: 'Grocery', value: 0.1, color: '#00A86B', amount: '412' }
-];
 
-const ProgressBar = () => {
+const ProgressBar = ({ transactions, selectedType }) => {
+    const categories = transactions.reduce((acc, transaction) => {
+        const { category, amount, transactionType } = transaction;
+        if (!acc[category]) {
+            acc[category] = { amount: 0, color: transactionType === 'income' ? '#00A86B' : '#FD3C4A' };
+        }
+        acc[category].amount += parseFloat(amount);
+        return acc;
+    }, {});
+
+    const totalAmount = Object.values(categories).reduce((acc, category) => acc + category.amount, 0);
     return (
         <ScrollView style={styles.container}>
-            {categories.map((category, index) => (
+            {Object.keys(categories).map((category, index) => (
                 <View key={index} style={styles.progressBarContainer}>
                     <View style={styles.detailsContainer}>
                         <View style={styles.labelContainer}>
-                            <View style={[styles.circle, { backgroundColor: category.color }]} />
-                            <Text style={styles.label}>{category.label}</Text>
+                            <View style={[styles.circle, { backgroundColor: categories[category].color }]} />
+                            <Text style={styles.label}>{category}</Text>
                         </View>
-                        <Text style={styles.amount}>- {category.amount}</Text>
+                        <Text
+                            style={[
+                                styles.amount,
+                                selectedType === 'income' ? styles.incomeText : styles.expenseText
+                            ]}
+                        >
+                            {selectedType === 'income' ? `+ ₹${categories[category].amount.toFixed(2)}` : `- ₹${categories[category].amount.toFixed(2)}`}
+                        </Text>
                     </View>
                     <Progress.Bar
-                        progress={category.value}
+                        progress={categories[category].amount / totalAmount}
                         width={Dimensions.get('window').width - 40}
-                        color={category.color}
+                        color={categories[category].color}
                         unfilledColor="#F1F1FA"
                         borderWidth={0}
                         height={12}
@@ -37,47 +47,5 @@ const ProgressBar = () => {
         </ScrollView>
     );
 };
-const styles = StyleSheet.create({
-    container: {
-        flexGrow: 1,
-        marginBottom: 20,
-        backgroundColor: '#FFF6E5'
-    },
-    progressBarContainer: {
-        flexDirection: 'column',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        marginBottom: 20,
-        height: 50,
-    },
-    detailsContainer: {
-        flexDirection: 'row',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-        width: Dimensions.get('window').width - 40,
-    },
-    labelContainer: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        alignItems: 'flex-start',
-        width: Dimensions.get('window').width * 0.3,
-    },
-    circle: {
-        width: 10,
-        height: 10,
-        borderRadius: 5,
-        marginRight: 10,
-        marginTop: 6
-    },
-    label: {
-        fontSize: 16,
-        color: '#212325'
-    },
-    amount: {
-        fontSize: 24,
-        color: '#FD3C4A',
-        marginLeft: 50,
-        alignSelf: 'flex-end'
-    }
-});
+
 export default ProgressBar
