+++ src/constant.js
@@ -1,0 +1,38 @@
+export const daysOfWeek = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
+export const monthsOfYear = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"];
+export const tabdata = ["Today", "Week", "Month", "Year"];
+
+export const categoryList = [
+    { label: 'Shopping', value: 'shopping' },
+    { label: 'Salary', value: 'salary' },
+    { label: 'Food', value: 'food' },
+    { label: 'Outing', value: 'outing' }
+];
+
+export const descriptionList = [
+    { label: 'Buy some grocery', value: 'Buy some grocery' },
+    { label: 'Arabian hut', value: 'Arabian hut' },
+    { label: 'Salary for August', value: 'Salary for August' },
+    { label: 'Outing', value: 'outing' }
+];
+
+export const months = [
+    { label: 'January', value: '1' },
+    { label: 'February', value: '2' },
+    { label: 'March', value: '3' },
+    { label: 'April', value: '4' },
+    { label: 'May', value: '5' },
+    { label: 'June', value: '6' },
+    { label: 'July', value: '7' },
+    { label: 'August', value: '8' },
+    { label: 'September', value: '9' },
+    { label: 'October', value: '10' },
+    { label: 'November', value: '11' },
+    { label: 'December', value: '12' },
+];
+
+export const categories = [
+    { label: 'All', value: '1' },
+    { label: 'Income', value: '2' },
+    { label: 'Expense', value: '3' },
+  ];
