+++ src/redux/action_type.js
@@ -1,0 +1,3 @@
+export const ADD_TRANSACTION = 'ADD_TRANSACTION';
+export const SELECTED_MONTH = 'SELECTED_MONTH';
+export const SELECTED_CATEGORY = 'SELECTED_CATEGORY';
