+++ src/screens/Profile.js
@@ -1,38 +1,40 @@
 import { Image, StyleSheet, Text, TouchableOpacity, View } from 'react-native';
 import React from 'react';
+import { accountImage, appLogo, editSymbol, exportDataImage, logoutImage, settingImage } from '../utils/images';
+import styles from '../styles/ProfileStyle';
 
 const Profile = () => {
   return (
     <View style={styles.container}>
       <View style={styles.headerContainer}>
         <View style={styles.imageContainer}>
-          <Image style={styles.profileImage} source={require('../../assets/images/applogo.png')} />
+          <Image style={styles.profileImage} source={appLogo} />
         </View>
         <View style={styles.userInfo}>
           <Text style={styles.usernameText}>Username</Text>
           <Text style={styles.nameText}>Sarthak Srivastava</Text>
         </View>
-        <Image source={require('../../assets/images/edit.png')} />
+        <Image source={editSymbol} />
       </View>
 
       <View style={styles.optionsContainer}>
         <TouchableOpacity style={styles.option}>
-          <Image source={require('../../assets/images/account.png')} />
+          <Image source={accountImage} />
           <Text style={styles.optionText}>Account</Text>
         </TouchableOpacity>
         <View style={styles.divider}></View>
         <TouchableOpacity style={styles.option}>
-          <Image source={require('../../assets/images/setting.png')} />
+          <Image source={settingImage} />
           <Text style={styles.optionText}>Settings</Text>
         </TouchableOpacity>
         <View style={styles.divider}></View>
         <TouchableOpacity style={styles.option}>
-          <Image source={require('../../assets/images/exportData.png')} />
+          <Image source={exportDataImage} />
           <Text style={styles.optionText}>Export Data</Text>
         </TouchableOpacity>
         <View style={styles.divider}></View>
         <TouchableOpacity style={styles.option}>
-          <Image source={require('../../assets/images/logout.png')} />
+          <Image source={logoutImage} />
           <Text style={styles.optionText}>Logout</Text>
         </TouchableOpacity>
       </View>
@@ -42,65 +44,4 @@ const Profile = () => {
 
 export default Profile;
 
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: '#FFF6E5',
-    flex: 1,
-  },
-  headerContainer: {
-    marginVertical: 100,
-    marginHorizontal: 20,
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-  },
-  imageContainer: {
-    width: 90,
-    height: 90,
-    borderRadius: 45,
-    borderColor: 'blue',
-    borderWidth: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    margin: 10,
-  },
-  profileImage: {
-    width: 80,
-    height: 80,
-    borderRadius: 40,
-    alignSelf: 'center',
-    margin: 2,
-  },
-  userInfo: {
-    alignSelf: 'center',
-  },
-  usernameText: {
-    fontSize: 14,
-  },
-  nameText: {
-    fontSize: 24,
-    fontWeight: '600',
-    color: 'black',
-  },
-  optionsContainer: {
-    backgroundColor: 'white',
-    flex: 0.9,
-    marginHorizontal: 10,
-    borderRadius: 24,
-  },
-  option: {
-    margin: 20,
-    flexDirection: 'row',
-  },
-  optionText: {
-    textAlignVertical: 'center',
-    marginLeft: 10,
-    fontSize: 16,
-    color: 'black',
-    fontWeight: '500',
-  },
-  divider: {
-    borderBottomColor: '#D3D3D3',
-    borderBottomWidth: StyleSheet.hairlineWidth,
-    marginHorizontal: 10,
-  },
-});
+
