+++ src/screens/Transaction.js
@@ -2,26 +2,8 @@ import { FlatList, Image, StyleSheet, Text, TouchableOpacity, View } from 'react
 import React, { useState } from 'react';
 import { Dropdown } from 'react-native-element-dropdown';
 import { useDispatch, useSelector } from 'react-redux';
-const months = [
-  { label: 'January', value: '1' },
-  { label: 'February', value: '2' },
-  { label: 'March', value: '3' },
-  { label: 'April', value: '4' },
-  { label: 'May', value: '5' },
-  { label: 'June', value: '6' },
-  { label: 'July', value: '7' },
-  { label: 'August', value: '8' },
-  { label: 'September', value: '9' },
-  { label: 'October', value: '10' },
-  { label: 'November', value: '11' },
-  { label: 'December', value: '12' },
-];
-
-const categories = [
-  { label: 'All', value: '1' },
-  { label: 'Income', value: '2' },
-  { label: 'Expense', value: '3' },
-];
+import styles from '../styles/TransactionStyle';
+import { months, categories } from '../constant';
 
 const Transaction = () => {
   const transactions = useSelector(state => state.transactionsReducer.transactions);
@@ -41,6 +23,7 @@ const Transaction = () => {
         transactions={transactions}
         selectedMonth={selectedMonth}
         selectedCategory={selectedCategory}
+        selectedDate={transactions.date}
 
       />
     </View>
@@ -91,7 +74,7 @@ const FilterComponent = ({ selectedMonth, selectedCategory }) => {
     </View>
   );
 };
-const Filterresult = ({ transactions, selectedMonth, selectedCategory }) => {
+const Filterresult = ({ transactions, selectedMonth, selectedCategory,selectedDate }) => {
   // Convert selectedMonth label to corresponding number
   const selectedMonthValue = months.find(month => month.label === selectedMonth)?.value;
 
@@ -112,8 +95,8 @@ const Filterresult = ({ transactions, selectedMonth, selectedCategory }) => {
   console.log({ selectedMonth });
   console.log({ selectedCategory });
 
-  const formatTime = (timestamp) => {
-    const date = new Date(timestamp);
+  const formatTime = (selectedDate) => {
+    const date = new Date(selectedDate);
     return date.toLocaleTimeString('en-US', {
       hour: 'numeric',
       minute: 'numeric',
@@ -143,7 +126,7 @@ const Filterresult = ({ transactions, selectedMonth, selectedCategory }) => {
                 {item.transactionType === 'expense' ? '-' : '+'}
                 {item.amount}
               </Text>
-              <Text style={styles.time}>{formatTime(item.timestamp)}</Text>
+              <Text style={styles.time}>{formatTime(item.date)}</Text>
             </View>
           </View>
         )}
@@ -151,64 +134,6 @@ const Filterresult = ({ transactions, selectedMonth, selectedCategory }) => {
     </View>
   );
 };
-
-
 export default Transaction;
 
-const styles = StyleSheet.create({
-  container: {
-    backgroundColor: '#FFF6E5',
-    flex: 1,
-  },
-  filterContainer: {
-    flexDirection: 'row',
-    marginTop: 40,
-  },
-  dropdown: {
-    paddingHorizontal: 16,
-    paddingVertical: 8,
-    flexDirection: 'row',
-    marginHorizontal: 10,
-    borderRadius: 40,
-    borderWidth: 1,
-    alignItems: 'center',
-    width: 130,
-    height: 40,
-  },
-  flatListContent: {
-    paddingBottom: 50,
-  },
-  transactionContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    height: 90,
-    backgroundColor: 'white',
-    margin: 10,
-    borderRadius: 24,
-  },
-  transactionDetails: {
-    flexDirection: 'column',
-    margin: 20,
-    justifyContent: 'space-between',
-  },
-  category: {
-    fontSize: 22,
-  },
-  description: {
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  transactionAmount: {
-    flexDirection: 'column',
-    justifyContent: 'space-between',
-    margin: 20,
-  },
-  amount: {
-    fontSize: 20,
-    fontWeight: '600',
-  },
-  time: {
-    fontSize: 13,
-    fontWeight: '600',
-  },
-});
+
