+++ src/utils/images.js
@@ -1,0 +1,11 @@
+export const appLogo = require('../../assets/images/applogo.png')
+export const incomeImage = require('../../assets/images/incomeArrow.png')
+export const cameraImage = require('../../assets/images/camera.png')
+export const expenseImage = require('../../assets/images/expense.png')
+export const ruppeSymbol = require('../../assets/images/bi_currency-rupee.png')
+export const editSymbol = require('../../assets/images/edit.png')
+export const accountImage = require('../../assets/images/account.png')
+export const settingImage = require('../../assets/images/setting.png')
+export const exportDataImage = require('../../assets/images/exportData.png')
+export const logoutImage = require('../../assets/images/logout.png')
+export const arrowDown = require('../../assets/images/arrow-down-2.png')
